[common] refactor (#1) camel case로 경로 변경
