+++ client/src/stories/Map.stories.tsx
