--- server/utils/authenticateNotRequired.ts
@@ -1,3 +1,0 @@
-const authenticateNotRequired = ['UserSignin', 'UserSignup', 'DriverSignin', 'DriverSignup'];
-
-export default authenticateNotRequired;
